Move UriImage to core entities
