+++ coreentities/src/main/java/com/neaniesoft/vermilion/coreentities/UriImage.kt
@@ -1,4 +1,4 @@
-package com.neaniesoft.vermilion.posts.domain.entities
+package com.neaniesoft.vermilion.coreentities
 
 import android.net.Uri
