+++ posts/src/main/java/com/neaniesoft/vermilion/posts/domain/ImageSizeChooser.kt
@@ -1,7 +1,7 @@
 package com.neaniesoft.vermilion.posts.domain
 
 import com.neaniesoft.vermilion.posts.domain.entities.AnimatedImagePreview
-import com.neaniesoft.vermilion.posts.domain.entities.UriImage
+import com.neaniesoft.vermilion.coreentities.UriImage
 
 fun List<UriImage>.choosePreviewImage(): UriImage? = maxByOrNull { it.width }
